+++ CHANGES.md
@@ -1,4 +1,11 @@
 # Tango Admin Change Log
 
+## 0.2 -
+* Lazy first effort at admin theme, but it's responsive.
+* Admin now has app menu on left side. Thanks to @kezabelle for the idea.
+* Moved inline link lists in from tango-shared
+* Moved text field counter in from tango-shared
+* Drag and drop tabular inlines
+
 ## 0.1
 * Initial release. Includes a "Nuke users" action, and an admin text counter widget.